requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
web3>=6.0.0
py-clob-client>=0.24.0
fastapi>=0.110.0
uvicorn>=0.27.0
